rename mesmer_x module to distrib
